mobile responsive
made the site fully responsive to all relevant sizes (i think), fixed the project images to fill the size of the div, made the form input boxes not a fixed width to fit into smaller view ports, removes some repeating information in the header section, fixed the li's kind of. last thing i need to do now is to build a server and host the site!